ai-worker/pyproject.toml
[project]
name = "ai-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.3.0",
    "confluent-kafka>=2.3.0",
    "pydantic>=2.5.0",
    "langchain>=1.2.15",
    "numpy>=2.4.4",
    "google-genai>=1.0.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "redis>=7.4.0",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "torch>=2.11.0",
    "transformers>=5.5.0",
    "python-dotenv>=1.2.2",
]
